Move buttons into their own directory and fix ssh username setting
