+++ frontend/src/buttons/BrowserButton/BrowserButton.tsx
@@ -3,7 +3,7 @@ import { hostName } from '../../helpers/nameHelper'
 import { IService } from 'remote.it'
 import { IconButton, Tooltip } from '@material-ui/core'
 import { makeStyles } from '@material-ui/styles'
-import { Icon } from '../Icon'
+import { Icon } from '../../components/Icon'
 
 type Props = {
   connection?: IConnection
@@ -16,7 +16,7 @@ export const BrowserButton: React.FC<Props> = ({ connection, service }) => {
   if (!connection || !connection.active) return null
 
   return (
-    <Tooltip title="Launch">
+    <Tooltip title="Launch Browser">
       <IconButton onClick={() => window.open(`http://${hostName(connection)}`)}>
         <Icon className={css.rotate} name="arrow-right" size="md" fixedWidth />
       </IconButton>
