+++ frontend/src/buttons/CopyButton/index.ts
