+++ frontend/src/components/PortSetting/PortSetting.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react'
 import BackendAdaptor from '../../services/BackendAdapter'
 import { IService } from 'remote.it'
 import { useSelector } from 'react-redux'
-import { ResetButton } from '../ResetButton'
+import { ResetButton } from '../../buttons/ResetButton'
 import { InlineSetting } from '../InlineSetting'
 import { REGEX_PORT_SAFE } from '../../constants'
 import { ApplicationState } from '../../store'
