+++ frontend/src/components/ServiceListItem/ServiceListItem.tsx
@@ -6,14 +6,14 @@ import { ConnectionStateIcon } from '../ConnectionStateIcon'
 import { ListItemIcon, ListItemText, ListItemSecondaryAction } from '@material-ui/core'
 import { ConnectionErrorMessage } from '../ConnectionErrorMessage'
 import { ListItemLocation } from '../ListItemLocation'
-import { DisconnectButton } from '../DisconnectButton'
-import { ConnectButton } from '../ConnectButton'
-import { BrowserButton } from '../BrowserButton'
+import { DisconnectButton } from '../../buttons/DisconnectButton'
+import { ConnectButton } from '../../buttons/ConnectButton'
+import { BrowserButton } from '../../buttons/BrowserButton'
 import { ServiceName } from '../ServiceName'
-import { CopyButton } from '../CopyButton'
+import { CopyButton } from '../../buttons/CopyButton'
 import { Throughput } from '../Throughput'
 import { makeStyles } from '@material-ui/styles'
-import { SSHButton } from '../SSHButton'
+import { SSHButton } from '../../buttons/SSHButton'
 
 export interface ServiceListItemProps {
   connection?: IConnection
