+++ frontend/src/pages/ServicePage/ServicePage.tsx
@@ -14,14 +14,14 @@ import { ApplicationState } from '../../store'
 import { Typography, Divider, List } from '@material-ui/core'
 import { ConnectionErrorMessage } from '../../components/ConnectionErrorMessage'
 import { ConnectionStateIcon } from '../../components/ConnectionStateIcon'
-import { DisconnectButton } from '../../components/DisconnectButton'
+import { DisconnectButton } from '../../buttons/DisconnectButton'
 import { LanShareSelect } from '../../components/LanShareSelect'
-import { ConnectButton } from '../../components/ConnectButton'
-import { BrowserButton } from '../../components/BrowserButton'
-import { ForgetButton } from '../../components/ForgetButton'
+import { ConnectButton } from '../../buttons/ConnectButton'
+import { BrowserButton } from '../../buttons/BrowserButton'
+import { ForgetButton } from '../../buttons/ForgetButton'
 import { DataDisplay } from '../../components/DataDisplay'
-import { CopyButton } from '../../components/CopyButton'
-import { SSHButton } from '../../components/SSHButton'
+import { CopyButton } from '../../buttons/CopyButton'
+import { SSHButton } from '../../buttons/SSHButton'
 import { Container } from '../../components/Container'
 import { Columns } from '../../components/Columns'
